fix(query-api): scan ClickHouse aggregate results at their real UInt64 type
sankey and investmentflow read several sum()/countIf()/countDistinctIf()
aggregates over UInt32-width columns and scan the results into Go
destinations narrower than what ClickHouse actually returns -- the
driver promotes every one of those aggregates to UInt64 regardless of
the source column's own width, and refuses to scan that into *float64
or *int64.

Per field, the fix follows what the reference plane's service layer
does to the value before it reaches the wire: fields it wraps in
float(...) (new_items, new_bugs, items_touched, churn, canceled_items)
now CAST to Float64 in SQL, keeping a single float64 path end to end;
fields it wraps in int(...) (missing_repo, missing_team) now scan into
uint64, matching the driver's actual result type directly.

Every fixture-backed test in both packages answers a Scan call by
reflecting on the Go destination type it is handed rather than
replaying the real clickhouse-go driver's own conversion rules, so
none of them exercised this path; each package gains a real-ClickHouse
(testcontainers) recurrence guard that scans these exact aggregate
shapes against a seeded schema, confirmed RED on the prior scan types
and GREEN on the fix. The CI integration-package inventory and shard
manifest are updated for both packages' first -tags=integration file.

Every 503 degradation site already logs its wrapped cause with route,
org id and request id; that telemetry is unchanged.

Ticket: CHAOS-5866

+++ cmd/query-api/internal/sankey/builders.go
@@ -96,7 +96,7 @@ func buildExpenseFlow(ctx context.Context, client QueryClient, startDay, endDay 
 
 	unplanned := maxFloat(0, newBugs)
 	rework := maxFloat(0, minFloat(unplanned, bugCompleted))
-	abandoned := maxFloat(0, minFloat(rework, float64(canceledItems)))
+	abandoned := maxFloat(0, minFloat(rework, canceledItems))
 
 	nodes := newNodeAccumulator()
 	edges := newEdgeAccumulator()
